refact:props로 전달되는것 zustand로 상태관리
