+++ src/components/PlayerCard/PlayerCard.tsx
@@ -2,18 +2,7 @@
 
 import React from "react";
 import Image from "next/image";
-
-interface PlayerCardProps {
-  name: string;
-  imageUrl: string;
-  number: string;
-  position: {
-    top: string;
-    left: string;
-  };
-  role: string;
-  rating: number;
-}
+import Player from "@/store/Today-player";
 
 const getColorClass = (rating: number) => {
   if (rating < 3.0) {
@@ -27,15 +16,32 @@ const getColorClass = (rating: number) => {
   }
 };
 
-const PlayerCard: React.FC<PlayerCardProps> = ({ name, imageUrl, number, position, role, rating }) => {
+const PlayerCard: React.FC<Player> = ({
+  name,
+  imageUrl,
+  number,
+  position,
+  role,
+  rating,
+  onClick,
+}) => {
   return (
-    <div className="relative my-4 rounded-custom border border-black p-1 bg-black w-122 h-166.87">
+    <div
+      onClick={onClick}
+      className="relative my-4 rounded-custom border border-black p-1 bg-black w-122 h-166.87"
+    >
       <div className="relative">
         <Image src={imageUrl} alt={name} width={113} height={151} />
-        <div className={`absolute px-1 rounded-custom text-lg ${getColorClass(rating)}`} style={{ top: "-13%", left: "88%" }}>
+        <div
+          className={`absolute px-1 rounded-custom text-lg ${getColorClass(rating)}`}
+          style={{ top: "-13%", left: "88%" }}
+        >
           {rating}
         </div>
-        <div className="bg-black text-white text-sm absolute px-1 " style={{ top: "0%", left: "0%" }}>
+        <div
+          className="bg-black text-white text-sm absolute px-1 "
+          style={{ top: "0%", left: "0%" }}
+        >
           {role}
         </div>
       </div>
@@ -44,7 +50,14 @@ const PlayerCard: React.FC<PlayerCardProps> = ({ name, imageUrl, number, positio
 
         <div className="ml-4 ">
           {name}
-          <Image src="/images/logo.svg" alt="로고다" width={41.37} height={44.17} className="absolute" style={{ top: "12%", left: "73%" }} />
+          <Image
+            src="/images/logo.svg"
+            alt="로고다"
+            width={41.37}
+            height={44.17}
+            className="absolute"
+            style={{ top: "12%", left: "73%" }}
+          />
         </div>
       </div>
     </div>
